moved NbnLocationsApp class
