+++ src/main/java/io/swagger/api/impl/NbnApiServiceImpl.java
@@ -1,7 +1,6 @@
 package io.swagger.api.impl;
 
 import io.swagger.api.NbnApiService;
-import io.swagger.api.NbnLocationApp;
 import io.swagger.api.NotFoundException;
 import io.swagger.api.impl.response.OperationResult;
 import io.swagger.model.NbnLocationsObject;
